Propuesta de restructura del proyecto
