+++ src/Index.scss
@@ -2,10 +2,10 @@
 
 @import './scss/mixins';
 @import './scss/global';
-@import './scss/navbar';
-@import './scss/files';
-@import './scss/questions';
-@import './scss/card';
+// @import './scss/navbar';
+// @import './scss/files';
+// @import './scss/questions';
+// @import './scss/card';
 
 // Posestratificación
-@import './scss/posestratificación/global-pos';
+// @import './scss/posestratificación/global-pos';
