+++ src/components/Navbar.jsx
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom';
+import './Navbar.scss';
 
 const Navbar = () => {
   return (
