+++ src/components/Navbar.scss
@@ -1,3 +1,5 @@
+@import '../Index.scss';
+
 .Navbar {
   width: 100%;
   height: 89px;
