+++ src/components/posestratificacion/Data.scss
@@ -1,3 +1,5 @@
+@import '../../Index.scss';
+
 .Data {
   @include adjust;
   @include between;
