+++ src/components/posestratificacion/Files.scss
@@ -1,4 +1,6 @@
-.Files {
+@import '../../Index.scss';
+
+.FilesPos {
   @include centrar;
   @include ajustar;
   justify-content: space-between;
@@ -44,6 +46,35 @@
   }
 
   .Upload-files {
+    .icon {
+      font-size: xx-large;
+      cursor: help;
+      color: gray;
+    }
+
+    .sizedBox {
+      width: 35px;
+    }
+
+    .Tip {
+      margin-right: 10px;
+
+      div {
+        width: 80px;
+        @include between;
+      }
+
+      .icon {
+        font-size: smaller;
+        color: $color-warning;
+      }
+
+      .iconCheck {
+        font-size: smaller;
+        color: $color-done;
+      }
+    }
+
     .isLoading {
       @include centrar;
 
@@ -80,15 +111,6 @@
       .Button:hover {
         opacity: 0.8;
         cursor: pointer;
-      }
-
-      .document {
-        display: grid;
-        grid-template-columns: repeat(2, 1fr);
-        gap: 10px;
-        font-size: 40px;
-        font-weight: 600;
-        width: 70px;
       }
     }
   }
