+++ src/components/visualizacion/Card.scss
@@ -1,3 +1,5 @@
+@import '../../Index.scss';
+
 .Card {
   margin-right: 20px;
   .Statement {
@@ -92,7 +94,19 @@
         }
       }
 
-      @import './color-picker';
+      .Color-picker {
+        position: absolute;
+        left: 180px;
+        .Button-color {
+          background: white;
+          @include centrar;
+          border-radius: 2px;
+          cursor: pointer;
+          border: 1px solid rgb(0, 0, 0, 0.3);
+          margin-top: 2px;
+          font-size: small;
+        }
+      }
     }
 
     .edit-answer {
