+++ src/components/visualizacion/Questions.scss
@@ -1,3 +1,5 @@
+@import '../../Index.scss';
+
 .Questions {
   @include ajustar;
   @include between;
