--- src/scss/posestratificación/_global-pos.scss
@@ -1,2 +1,0 @@
-@import './filesPos';
-@import './data';
